requirements.txt
flask
torch
transformers
hydra-core
omegaconf
tqdm
python-Levenshtein